(NPUP-15) Port multiply operator to the dispatcher.
Porting the multiply operator's implementation to the dispatcher.
