+++ lib/src/compiler/evaluation/dispatcher.cc
@@ -40,6 +40,7 @@
 #include <puppet/compiler/evaluation/operators/binary/match.hpp>
 #include <puppet/compiler/evaluation/operators/binary/minus.hpp>
 #include <puppet/compiler/evaluation/operators/binary/modulo.hpp>
+#include <puppet/compiler/evaluation/operators/binary/multiply.hpp>
 #include <puppet/compiler/evaluation/operators/binary/not_match.hpp>
 #include <puppet/compiler/evaluation/operators/unary/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/unary/logical_not.hpp>
@@ -96,6 +97,7 @@ namespace puppet { namespace compiler { namespace evaluation {
         add(binary::match::create_descriptor());
         add(binary::minus::create_descriptor());
         add(binary::modulo::create_descriptor());
+        add(binary::multiply::create_descriptor());
         add(binary::not_match::create_descriptor());
 
         // Add the built-in unary operators
